## Description

<!--
Include reference to internal ticket "Fixes WB-NNNNN" and/or GitHub issue "Fixes #NNNN" (if applicable)
-->

This pr implements a query optimization using uuid7s, which are already indexed, as a pre-groupby datetime filter. This applies to all calls query requests (`calls_query_stats`, `calls_query`, `calls_query_stream`) that include a datetime filter. We now convert the timestamps provided into their equivalent uuid7s, which are indexed and sorted, massively decreasing the amount of granules that must be read into memory. This can occur before the groupby, because both call starts and ends have an id (the same id). 

Example stats query in prod: 
```sql
SELECT count() as c
FROM
(
    SELECT calls_merged.id AS id
    FROM calls_merged
    WHERE calls_merged.project_id = 'project_id'
    GROUP BY (calls_merged.project_id, calls_merged.id)
    HAVING (any(calls_merged.started_at) > _CAST(1742472000, 'UInt64')) AND (NOT (any(calls_merged.started_at) > _CAST(1742475600, 'UInt64'))) AND (any(calls_merged.deleted_at) IS NULL) AND (NOT (any(calls_merged.started_at) IS NULL))
)
```

Example stats query with optimization:
```sql
SELECT count() as c
FROM
(
    SELECT calls_merged.id AS id
    FROM calls_merged
    WHERE calls_merged.project_id = 'project_id'
        AND calls_merged.id > '0195b36c-b200-7000-8000-000000000000' AND calls_merged.id < '0195b3a3-a080-7000-8000-000000000000'
    GROUP BY (calls_merged.project_id, calls_merged.id)
    HAVING (any(calls_merged.started_at) > _CAST(1742472000, 'UInt64')) AND (NOT (any(calls_merged.started_at) > _CAST(1742475600, 'UInt64'))) AND (any(calls_merged.deleted_at) IS NULL) AND (NOT (any(calls_merged.started_at) IS NULL))
)
```

TODO: 
- can we construct a case where this id optimization fails? 
   - yes, when the ids are not uuid7. We fail miserably in this case. 

## Testing

add unit tests

qa-azure: 

prod stats query in big project with filter: [10 seconds](https://us5.datadoghq.com/apm/traces?query=env%3Amanaged-install%20operation_name%3Afastapi.request%20service%3Aweave-trace%20%40weave_trace_server.auth_scope.email%3Agriffin.tarpenning%40wandb.com&agg_m=count&agg_m_source=base&agg_t=count&cols=service%2Cresource_name%2C%40duration%2C%40http.method%2C%40http.status_code%2C%40_span.count%2C%40_duration.by_service%2C%40weave_trace_server.auth_scope.email&fromUser=false&graphType=flamegraph&historicalData=false&messageDisplay=inline&query_translation_version=v0&refresh_mode=sliding&shouldShowLegend=true&sort=time&spanID=9852347362044463913&spanType=all&spanViewType=logs&storage=hot&timeHint=1743113551411&trace=67e5cd44000000002fed8ea33b1a5c919852347362044463913&traceID=67e5cd44000000002fed8ea33b1a5c91&view=spans&start=1743112662057&end=1743113562057&paused=false)

branch stats query in big project with filter: [650 ms](https://us5.datadoghq.com/apm/traces?query=env%3Amanaged-install%20operation_name%3Afastapi.request%20service%3Aweave-trace%20%40weave_trace_server.auth_scope.email%3Agriffin.tarpenning%40wandb.com%20-resource_name%3A%22POST%20%2Ftraces%2Ffeedback%2Fquery%22&agg_m=count&agg_m_source=base&agg_t=count&cols=service%2Cresource_name%2C%40duration%2C%40http.method%2C%40http.status_code%2C%40_span.count%2C%40_duration.by_service%2C%40weave_trace_server.auth_scope.email&fromUser=false&graphType=flamegraph&historicalData=false&messageDisplay=inline&query_translation_version=v0&refresh_mode=sliding&shouldShowLegend=true&sort=time&spanID=14711125143075388032&spanType=all&spanViewType=metadata&storage=hot&timeHint=1743113262593&trace=67e5cc2d000000005d5baad8d59273f414711125143075388032&traceID=67e5cc2d000000005d5baad8d59273f4&view=spans&start=1743112740300&end=1743113640300&paused=false) 

Can be even better. In a specific high load dedicated server environment, string filter counts: 

| stat      | prod      | branch   |
|-----------|-----------|----------|
| mem       | 17GB      | 91MB     |
| rows read | 34,673,371| 76,744   |
| time      | 12.5 s    | 0.026 s  |