@gtarpenning 2025-04-03
Closing in favor of: https://github.com/wandb/weave/pull/4035